use explicit string literals for id/beacon in nearobject clauses
When the 'id' or 'beacon' value is not put between quotes and a
pure string is provided as value (e.g., 'test_id'), i.e., without
digits, the produced GraphQL query is fine. However, as soon as
UUIDs are provided, as is relevant for 'nearobject' clauses,
these UUIDs are interpreted as numbers, and it breaks after the
first dash.